refactor(analysis): migrate analysis modules
