+++ src/tilefoundry/analysis/poly/__init__.py
@@ -1,0 +1,27 @@
+"""Public polyhedral analysis API."""
+
+from .access import (
+    AccessFootprint,
+    AxisExtent,
+    access_footprints,
+    carried_distances,
+    statement_time_dims,
+    time_extents,
+)
+from .errors import ExtractError
+from .extract import _local_type, extract
+from .model import TileGraph, TileUnit
+
+__all__ = [
+    "AccessFootprint",
+    "AxisExtent",
+    "ExtractError",
+    "TileGraph",
+    "TileUnit",
+    "access_footprints",
+    "carried_distances",
+    "extract",
+    "_local_type",
+    "statement_time_dims",
+    "time_extents",
+]
