+++ src/tilefoundry/analysis/poly/errors.py
@@ -1,0 +1,8 @@
+"""Errors raised while extracting polyhedral analysis data."""
+
+
+class ExtractError(NotImplementedError):
+    """A construct ``extract`` does not (yet) support."""
+
+
+__all__ = ["ExtractError"]
